git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C26463

+++ CDRAnalysis/Vanrise.Fzero.FraudAnalysis.Entities/StrategyExecution.cs
@@ -10,7 +10,7 @@ namespace Vanrise.Fzero.FraudAnalysis.Entities
     {
         public int ID { get; set; }
 
-        public int ProcessID { get; set; }
+        public long ProcessID { get; set; }
 
         public int StrategyID { get; set; }
